requirements.txt
pypdf==4.3.1
cohere
tiktoken
langchain==0.2.14
sentence-transformers==3.0.1
openai==1.42.0
pinecone-client==5.0.1
langchain-community==0.2.12
langchain-pinecone==0.1.3
langchain_openai==0.1.22